change shortcut for 'split cell' from Alt+- to Alt+Enter
The problem with Alt+- is that it does not work with keyboard layouts
that require you to press alt in order to enter a minus sign (e.g. the
German Neo layout).  This leads to the split action firing every time
one tries to enter a minus (i.e., much too often).